Squash a few todos and upgrade methods
This brief release updates the README.md with some
objectives I wish to accomplish with the Main.java
file.
It also tests some of the methods to be used
(primarily, trecTopic.java, DocGraph.java and
more).
Finally, many TODOs are completed, in reality
there's only testing to be done of the file
tools methods as well as implementing Grid search
in a linear way (then testing it for running in
parallel way).